backend/requirements.txt
langchain
langchain-community
langchain-google-genai 
google-auth 
google-auth-oauthlib 
google-api-python-client 
python-dotenv
pydantic
pymongo[srv]
dnspython
